pyproject.toml
[project]
name = "my-helper-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Nour Daniil",email = "daniilnour@yandex.ru"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]

[tool.poetry]
packages = [{include = "my_helper_server", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
